+++ application/views/dashboard/penjualan/footer.php
@@ -6,18 +6,18 @@
 
     $(document).ready(function() {
         
-        // function selectKategori() {
-        //     var selectedKategori = document.getElementById('select-kategori');
-        //     $(selectedKategori).change(function() {
-        //         var kodeKat = this.options[this.selectedIndex].value;
-        //         $.ajax({
-        //             url : baseUrl+'barang/findMaxId/'+kodeKat,
-        //             dataType : 'json'
-        //         }).done(function(r){
-        //             $('[name=kode_brg]').val(r.kode_brg);
-        //         }); 
-        //     });
-        // }
+        function selectKategori() {
+            var selectedKategori = document.getElementById('select-kategori');
+            $(selectedKategori).change(function() {
+                var kodeKat = this.options[this.selectedIndex].value;
+                $.ajax({
+                    url : baseUrl+'barang/findMaxId/'+kodeKat,
+                    dataType : 'json'
+                }).done(function(r){
+                    $('[name=kode_brg]').val(r.kode_brg);
+                }); 
+            });
+        }
 
         function dataTable() {
             var tableCari = $('#cari_barang').DataTable({
@@ -29,15 +29,8 @@
                     var kode_brg = $(this).attr('kode-brg');
                     $('[name=kode_brg]').val(kode_brg);
                     $('#modal-data-barang').modal('hide');
-                    // cariBarang();
+                    cariBarang();
                 });
-            });
-        }
-
-        function autocomplete() {
-            $( "#barang" ).autocomplete({
-                source: baseUrl+'barang/autocomplete',
-                minLenght: 3
             });
         }
 
@@ -90,8 +83,7 @@
         }
 
         function kodeBarangChange() {
-            $("#barang").click(function(){
-                console.log("wohe");
+            $("#barang").change(function(){
                 cariBarang();
             });
         }
@@ -181,9 +173,6 @@
                 rightAlign: false,
                 oncleared: function () { self.Value(''); }
             });
-            if ($('#barang')) {
-                autocomplete();
-            }
             $('#modal-data-barang').on('hidden.bs.modal', function () {
               $(this).data('bs.modal', null);
             });